Rename Iceberg metadata modules and scan type
